Fix tests, addressed Javadoc feedback
